pyproject.toml
[project]
name = "jirahhh"
version = "0.1.0"
description = "Utilities for creating and managing Jira issues with proper formatting and markdown support"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "jira>=3.5.0",
    "pypandoc>=1.11",
    "pyyaml>=6.0",
]

[project.scripts]
jirahhh = "jirahhh.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["jirahhh"]
